+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Python context manager `with flushing(MainEngine()) as eng:`
+-   Added OpenQASMBackend to output QASM from ProjectQ circuits
+-   Added projectq.libs.qasm to convert QASM to ProjectQ circuits
 
 ### Fixed